pyproject.toml
[project]
name = "fastapi-supabase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "supabase>=2.10.0",
]

[dependency-groups]
dev = [
    "ipython>=8.29.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.17",
    "uvicorn>=0.32.0",
]

[tool.setuptools]
packages = ["fastapi_supabase"]
